Moved IHostSettingsService to DotNetNuke.Abstractions.Application
